refact: mainpage 분리
